[Apollo] A dedicated module for network partitioning tests
Extracting the existing network partitioning tests into a separate module. More to come as part of the edge scenarios...
