+++ tests/apollo/test_skvbc_view_change.py
@@ -124,47 +124,6 @@ class SkvbcViewChangeTest(unittest.TestCase):
     @with_trio
     @with_bft_network(start_replica_cmd)
     @verify_linearizability
-    async def test_single_vc_primary_isolated(self, bft_network, tracker):
-        """
-        The goal of this test is to check the view change
-        workflow in case the primary is up, but its outgoing
-        communication is intercepted by an adversary.
-
-        1) Given a BFT network,
-        2) Insert an adversary that isolates the primary's outgoing communication
-        3) Send a batch of write requests.
-        4) Verify the BFT network eventually transitions to the next view.
-        5) Perform a "read-your-writes" check in the new view
-        """
-        with net.PrimaryIsolatingAdversary(bft_network) as adversary:
-            bft_network.start_all_replicas()
-
-            initial_primary = 0
-            await bft_network.wait_for_view(
-                replica_id=initial_primary,
-                expected=lambda v: v == initial_primary,
-                err_msg="Make sure we are in the initial view "
-                        "before isolating the primary."
-            )
-
-            await adversary.interfere()
-            expected_next_primary = 1
-
-            await self._send_random_writes(tracker)
-
-            await bft_network.wait_for_view(
-                replica_id=random.choice(bft_network.all_replicas(without={0})),
-                expected=lambda v: v == expected_next_primary,
-                err_msg="Make sure view change has been triggered."
-            )
-
-            await self._wait_for_read_your_writes_success(tracker)
-
-            await tracker.run_concurrent_ops(100)
-
-    @with_trio
-    @with_bft_network(start_replica_cmd)
-    @verify_linearizability
     async def test_single_vc_with_f_replicas_down(self, bft_network, tracker):
         """
         Here we "step it up" a little bit, bringing down a total of f replicas
